📦 FEAT: move pages into pages directory
